changed int to float

+++ map.cpp
@@ -1,6 +1,6 @@
 #include "map.h"
 
-Vector::Vector(int x, int y, int z): x(x), y(y), z(z) {}
+Vector::Vector(float x, float y, float z): x(x), y(y), z(z) {}
 Vector::Vector(const Vector &v): x(v.x), y(v.y), z(v.y) {}
